chore(rollbar): move sdk loader to a separate file
